Revert "Feature/mainview"
