--- OneLabUnsplash/Others/Assets.xcassets/image.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "emma-harrisova-pJgWvQPdvvE-unsplash 2.jpg",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "emma-harrisova-pJgWvQPdvvE-unsplash 2-1.jpg",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "emma-harrisova-pJgWvQPdvvE-unsplash 2-2.jpg",
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
