--- OneLabUnsplash/Others/SceneDelegate.swift
@@ -1,43 +1,0 @@
-//
-//  SceneDelegate.swift
-//  OneLabUnsplash
-//
-//  Created by Alikhan Abutalipov on 1/14/21.
-//
-
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
-    var window: UIWindow?
-
-
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-
-        let window = UIWindow(windowScene: windowScene)
-        let viewModel = MainViewModel()
-        let mainViewController = MainViewController(viewModel: viewModel)
-        let navigationController = UINavigationController(rootViewController: mainViewController)
-        window.rootViewController = navigationController
-        window.makeKeyAndVisible()
-        self.window = window
-    }
-
-    func sceneDidDisconnect(_ scene: UIScene) {
-    }
-
-    func sceneDidBecomeActive(_ scene: UIScene) {
-    }
-
-    func sceneWillResignActive(_ scene: UIScene) {
-    }
-
-    func sceneWillEnterForeground(_ scene: UIScene) {
-    }
-
-    func sceneDidEnterBackground(_ scene: UIScene) {
-        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
-    }
-}
-
