--- OneLabUnsplash/Presentation Layer/StretchyHeaderView.swift
@@ -1,65 +1,0 @@
-//
-//  StretchyHeaderView.swift
-//  OneLabUnsplash
-//
-//  Created by Айдана on 1/31/21.
-//
-
-import UIKit
-
-final class StretchyHeaderView: UIView {
-    
-    let imageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.clipsToBounds = true
-        imageView.contentMode = .scaleAspectFill
-        return imageView
-    }()
-    
-    private var containerView = UIView()
-    
-    private var imageViewHeightConstraint = NSLayoutConstraint()
-    private var imageViewBottomConstraint = NSLayoutConstraint()
-    private var containerViewHeightConstraint = NSLayoutConstraint()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureViews()
-        configureViewConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func configureViews() {
-        addSubview(containerView)
-        containerView.addSubview(imageView)
-    }
-    
-    func configureViewConstraints() {
-        NSLayoutConstraint.activate([
-            widthAnchor.constraint(equalTo: containerView.widthAnchor),
-            centerXAnchor.constraint(equalTo: containerView.centerXAnchor),
-            heightAnchor.constraint(equalTo: containerView.heightAnchor)
-        ])
-       
-        containerView.translatesAutoresizingMaskIntoConstraints = false
-        containerView.widthAnchor.constraint(equalTo: imageView.widthAnchor).isActive = true
-        containerViewHeightConstraint = containerView.heightAnchor.constraint(equalTo: self.heightAnchor)
-        containerViewHeightConstraint.isActive = true
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageViewBottomConstraint = imageView.bottomAnchor.constraint(equalTo: containerView.bottomAnchor)
-        imageViewBottomConstraint.isActive = true
-        imageViewHeightConstraint = imageView.heightAnchor.constraint(equalTo: containerView.heightAnchor)
-        imageViewHeightConstraint.isActive = true
-    }
-    
-    func scrollViewDidScroll(scrollView: UIScrollView) {
-        containerViewHeightConstraint.constant = scrollView.contentInset.top
-        let offsetY = -(scrollView.contentOffset.y + scrollView.contentInset.top)
-        containerView.clipsToBounds = offsetY <= 0
-        imageViewBottomConstraint.constant = offsetY >= 0 ? 0 : -offsetY / 2
-        imageViewHeightConstraint.constant = max(offsetY + scrollView.contentInset.top, scrollView.contentInset.top)
-    }
-}
