+++ OneLabUnsplash/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  OneLabUnsplash
+//
+//  Created by Alikhan Abutalipov on 1/14/21.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
